feat: all services working with docker compose
