+++ orders_pubsub/.dockerignore
@@ -1,0 +1,2 @@
+.venv
+__py**
