+++ orders_pubsub/main.py
@@ -1,0 +1,3 @@
+import src.message_broker.consumer as consumer
+
+consumer.consume()
