+++ orders_pubsub/src/config/vars.py
@@ -1,0 +1,6 @@
+import os
+
+ENV = {
+  'kafka_host': os.getenv('KAFKA_HOST', 'localhost:9092'),
+  'db_host': os.getenv('DB_HOST', 'localhost'),
+}
