+++ orders_pubsub/src/db/__init__.py
@@ -1,0 +1,1 @@
+from .register_order import register_order
